+++ tests/api_v2/fixtures/initializers.h
@@ -228,6 +228,7 @@ namespace ariles_tests
 
             ~ROS2Initializer()
             {
+                /*
                 rcl_interfaces::msg::ListParametersResult parameters = nh_->list_parameters({}, 100);
                 std::cout << ">>>>>>>>>" << std::endl;
                 for (const std::string &name : parameters.names)
@@ -235,6 +236,7 @@ namespace ariles_tests
                     std::cout << name << " = " << nh_->get_parameter(name).value_to_string() << std::endl;
                 }
                 std::cout << ">>>>>>>>>" << std::endl;
+                */
                 rclcpp::shutdown();
             }